requirements.txt
torch==2.1.0
torch-geometric==2.4.0
torch-scatter==2.1.2
torch-sparse==0.6.18
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nibabel==5.1.0
nilearn==0.10.2
dipy==1.7.0
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0.1
tqdm==4.66.1
scipy==1.11.1
networkx==3.1
joblib==1.3.2
